Major changes to the core of the cog
Thanks to Sinbad for helping with these changes!
The config is now only written to disk every 10 minutes OR when a command is used. This should reduce the strain on servers with more messages per second. This may increase load time on certain commands, I may add a fix for this in the future.